Add files via upload
Objective 
Congratulations,We have landed on mars and we are trying to colonize it. But the main problem we are facing is which plants we can grow at different geographical regions so that maximum number of plants would survive. 
Fortunately we have the data from earth to give us insights. Please help us in predicting which kind of plant we can grow at different locations depending on the environment. 
It’s time to save humanity which is at the brink of extinction. 

Problem Statement
1. Create a notebook on key insights derived from Exploratory Data Analysis
2. Create a multi-class prediction model to predict the species of plant which will survive in the neighborhood of a given environment.
3. Create key segments for all the plants (train + test) based on the average sunlight 

received throughout the day and their distance from waterbody to identify which 

segments of plants are getting enough sunlight and water vs which ones are not. This will help in mobilizing resources to track growth of trees appropriately